requirements.txt
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
psycopg[binary]==3.2.13
alembic==1.16.4
pytest==8.4.1
